+++ scripts/pallet-weight-template.hbs
@@ -34,9 +34,6 @@ use frame_support::{traits::Get, weights::{Weight, constants::RocksDbWeight}};
 use core::marker::PhantomData;
 use crate::*;
 
-/// Weights for `{{pallet}}`.
-pub struct WeightInfo<T>(PhantomData<T>);
-
 /// Weights for `{{pallet}}` using the Neuroweb node and recommended hardware.
 pub struct NeurowebWeight<T>(PhantomData<T>);
 impl<T: frame_system::Config> {{pallet}}::WeightInfo for NeurowebWeight<T> {
